+++ geode-core/src/test/java/org/apache/geode/internal/cache/tier/sockets/command/TXSynchronizationCommandTest.java
@@ -26,7 +26,6 @@ import javax.transaction.Status;
 
 import org.junit.Before;
 import org.junit.Test;
-import org.junit.experimental.categories.Category;
 
 import org.apache.geode.cache.client.internal.TXSynchronizationOp;
 import org.apache.geode.distributed.internal.membership.InternalDistributedMember;
@@ -38,9 +37,7 @@ import org.apache.geode.internal.cache.tier.Command;
 import org.apache.geode.internal.cache.tier.sockets.Message;
 import org.apache.geode.internal.cache.tier.sockets.Part;
 import org.apache.geode.internal.cache.tier.sockets.ServerConnection;
-import org.apache.geode.test.junit.categories.UnitTest;
 
-@Category(UnitTest.class)
 public class TXSynchronizationCommandTest {
   private Message clientMessage;
   private ServerConnection serverConnection;
